Reinclusion of EC1, EC2 and EC4 onto EC3_dev.
The folders related to EC1, EC2 and EC4 were inadvertedly removed from the EC3_dev branch during the EC3 development by L. Garcia on Feb. 2023. Such a removal could have caused potential corruption on the GitHub repository when the EC3_dev branch was to be merged onto the 'main' branch after the finishing of the case study EC3.

As a result, this push aims to ammend the prior exclusion of the corresponding folders of EC1, EC2, and EC4. The most recent pushes of EC1 (main branch), EC2 (main branch), and EC4 (EC4_dev branch) were added to this commit onto the EC3_dev branch. As a result, once EC3_dev and main are merged, 'main' will feature the most recent versions of all case studies.